Aire39 Puzzle Scene Part 1 (#35)
* add new tilemap and scene

* add fix to test move script so the particles effects don't have to be attached for the script to run

added new textures to help build the tower

progress on building tower outside and inside. Also progess of building the outside area

* remove closed off door tiles from the outside since there are now tiles for the inside

disable collision for inside objects since this needs to be controlled through a script

* add script that allows entry and exit from the tower

* add more to inside the tower 1st floor

add trigger to release ceiling trap-door

* add the beginning of chain/rope construction needed for 2nd floor puzzle

add quick force passage for testing

add new tag layer for rope/chain

* complete basic pulley system code

test added to be able to left-click crank to add link and rikght-click to remove link

* add source of adapted chain/rope code

* create crank for pulley

fix layer order of tile grids and minor updates to Tower level

* make jump force configureable and set to 40

* update to use the MainCharacter prefab and GroundCollider prefab for the 1st level platforms

Added AudioManager

* make it easier to enable enter/exit of tower for testing

* added a way to set the raycastCircle position by an offset. The few scripts that used this needed to be updated and don't add an offset

for the tower level. I changed the way the detection was done by centering the raycasat to the center of the character instead which needed an offset for it to work

* add control flip for part of the level

remove need to add edge of camera since this should be done by the camera follow script

* add melty effect thbat only effeects the player when using the ice power

* add more platforms

code cleanup

* update blender animation/character file

* rename tower scene to level_tower

* fix unwanted pixels showing up in ball character animation

* fix more unneeded pixels with the other character animation sheets that would appear at the top of the character

made sure to set the tower backside platforms initially to the disable state

* trap door will work with the fire attack

* fix block passage script

add setttings prefab

add enemies

---------

Co-authored-by: Sean Butler <seanxbutler@gmail>